DBI methods

Implementations of pure virtual functions defined in the DBI package.

Usage

# S4 method for KazamConnection
dbAppendTable(conn, name, value, ..., row.names = NULL)

# S4 method for KazamConnection
dbBegin(conn, ...)

# S4 method for KazamResult
dbBind(res, params, ...)

# S4 method for KazamResult
dbClearResult(res, ...)

# S4 method for KazamResultArrow
dbClearResult(res, ...)

# S4 method for KazamResult
dbColumnInfo(res, ...)

# S4 method for KazamConnection
dbCommit(conn, ...)

# S4 method for KazamDriver
dbConnect(drv, ...)

# S4 method for KazamConnection,ANY
dbDataType(dbObj, obj, ...)

# S4 method for KazamDriver,ANY
dbDataType(dbObj, obj, ...)

# S4 method for KazamDriver,list
dbDataType(dbObj, obj, ...)

# S4 method for KazamConnection
dbDisconnect(conn, ...)

# S4 method for KazamConnection,character
dbExistsTable(conn, name, ...)

# S4 method for KazamResultArrow
dbFetchArrow(res, ...)

# S4 method for KazamResult
dbFetch(res, n = -1, ...)

# S4 method for KazamConnection
dbGetInfo(dbObj, ...)

# S4 method for KazamDriver
dbGetInfo(dbObj, ...)

# S4 method for KazamResult
dbGetInfo(dbObj, ...)

# S4 method for KazamResult
dbGetRowCount(res, ...)

# S4 method for KazamResult
dbGetRowsAffected(res, ...)

# S4 method for KazamResult
dbGetStatement(res, ...)

# S4 method for KazamResult
dbHasCompleted(res, ...)

# S4 method for KazamConnection
dbIsValid(dbObj, ...)

# S4 method for KazamDriver
dbIsValid(dbObj, ...)

# S4 method for KazamResult
dbIsValid(dbObj, ...)

# S4 method for KazamConnection,character
dbListFields(conn, name, ...)

# S4 method for KazamConnection
dbListTables(conn, ...)

# S4 method for KazamConnection,character
dbQuoteIdentifier(conn, x, ...)

# S4 method for KazamConnection,character
dbQuoteString(conn, x, ...)

# S4 method for KazamConnection,character
dbRemoveTable(conn, name, ..., temporary = FALSE, fail_if_missing = TRUE)

# S4 method for KazamConnection
dbRollback(conn, ...)

# S4 method for KazamConnection
dbSendQueryArrow(conn, statement, ..., params = NULL)

# S4 method for KazamConnection,character
dbSendQuery(conn, statement, ..., params = NULL)

# S4 method for KazamConnection,character
dbSendStatement(conn, statement, ..., params = NULL)

# S4 method for KazamConnection,character,data.frame
dbWriteTable(
  conn,
  name,
  value,
  overwrite = FALSE,
  append = FALSE,
  ...,
  field.types = NULL,
  row.names = NULL,
  temporary = FALSE
)

# S4 method for KazamConnection
show(object)

# S4 method for KazamDriver
show(object)

# S4 method for KazamResult
show(object)

Arguments

conn

A DBIConnection object, as returned by dbConnect().

name

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')

value

For dbWriteTable(), a data.frame (or coercible to data.frame). For dbWriteTableArrow(), an object coercible to an Arrow RecordBatchReader.

...

Other parameters passed on to methods.

row.names

A logical specifying whether the row.names should be output to the output DBMS table; if TRUE, an extra field whose name will be whatever the R identifier "row.names" maps to the DBMS (see DBI::make.db.names()). If NA will add rows names if they are characters, otherwise will ignore.

res

An object inheriting from DBIResult.

params

For dbBind(), a list of values, named or unnamed, or a data frame, with one element/column per query parameter. For dbBindArrow(), values as a nanoarrow stream, with one column per query parameter.

drv

an object that inherits from DBIDriver, or an existing DBIConnection object (in order to clone an existing connection).

dbObj

A object inheriting from DBIDriver or DBIConnection

obj

An R object whose SQL type we want to determine.

n

maximum number of records to retrieve per fetch. Use n = -1 or n = Inf to retrieve all pending records. Some implementations may recognize other special values.

x

A character vector, SQL or Id object to quote as identifier.

temporary

a logical specifying whether the new table should be temporary. Its default is FALSE.

fail_if_missing

If FALSE, dbRemoveTable() succeeds if the table doesn't exist.

statement

a character string containing SQL.

overwrite

Allow overwriting the destination table. Cannot be TRUE if append is also TRUE.

append

Allow appending to the destination table. Cannot be TRUE if overwrite is also TRUE.

field.types

character vector of named SQL field types where the names are the names of new table's columns. If missing, types inferred with DBI::dbDataType()).

object

Any R object