DBI methods

Implementations of pure virtual functions defined in the DBI package.

# S4 method for KazamDriver
show(object)

# S4 method for KazamDriver
dbConnect(drv, ...)

# S4 method for KazamDriver,ANY
dbDataType(dbObj, obj, ...)

# S4 method for KazamDriver,list
dbDataType(dbObj, obj, ...)

# S4 method for KazamDriver
dbIsValid(dbObj, ...)

# S4 method for KazamDriver
dbGetInfo(dbObj, ...)

# S4 method for KazamConnection
show(object)

# S4 method for KazamConnection
dbIsValid(dbObj, ...)

# S4 method for KazamConnection
dbDisconnect(conn, ...)

# S4 method for KazamConnection,character
dbSendQuery(conn, statement, ...)

# S4 method for KazamConnection,character
dbSendStatement(conn, statement, ...)

# S4 method for KazamConnection,ANY
dbDataType(dbObj, obj, ...)

# S4 method for KazamConnection,character
dbQuoteString(conn, x, ...)

# S4 method for KazamConnection,character
dbQuoteIdentifier(conn, x, ...)

# S4 method for KazamConnection,character,data.frame
dbWriteTable(conn, name, value, overwrite = FALSE, append = FALSE, ...)

# S4 method for KazamConnection,character
dbReadTable(conn, name, ...)

# S4 method for KazamConnection
dbListTables(conn, ...)

# S4 method for KazamConnection,character
dbExistsTable(conn, name, ...)

# S4 method for KazamConnection,character
dbListFields(conn, name, ...)

# S4 method for KazamConnection,character
dbRemoveTable(conn, name, ...)

# S4 method for KazamConnection
dbGetInfo(dbObj, ...)

# S4 method for KazamConnection
dbBegin(conn, ...)

# S4 method for KazamConnection
dbCommit(conn, ...)

# S4 method for KazamConnection
dbRollback(conn, ...)

# S4 method for KazamResult
show(object)

# S4 method for KazamResult
dbClearResult(res, ...)

# S4 method for KazamResult
dbFetch(res, n = -1, ...)

# S4 method for KazamResult
dbHasCompleted(res, ...)

# S4 method for KazamResult
dbGetInfo(dbObj, ...)

# S4 method for KazamResult
dbIsValid(dbObj, ...)

# S4 method for KazamResult
dbGetStatement(res, ...)

# S4 method for KazamResult
dbColumnInfo(res, ...)

# S4 method for KazamResult
dbGetRowCount(res, ...)

# S4 method for KazamResult
dbGetRowsAffected(res, ...)

# S4 method for KazamResult
dbBind(res, params, ...)

Arguments

object
Any R object

drv
an object that inherits from DBIDriver, or an existing DBIConnection object (in order to clone an existing connection).

...
authentication arguments needed by the DBMS instance; these typically include user, password, host, port, dbname, etc. For details see the appropriate DBIDriver.

dbObj
A object inheriting from DBIDriver or DBIConnection

obj
An R object whose SQL type we want to determine.

conn
A DBIConnection object, as returned by dbConnect().

statement
a character string containing SQL.

x
A character vector to quote as string.

name
A character string specifying the unquoted DBMS table name, or the result of a call to dbQuoteIdentifier().

value
a data.frame (or coercible to data.frame).

overwrite
Allow overwriting the destination table. Cannot be TRUE if append is also TRUE.

append
Allow appending to the destination table. Cannot be TRUE if overwrite is also TRUE.

res
An object inheriting from DBIResult.

n
maximum number of records to retrieve per fetch. Use n = -1 or n = Inf to retrieve all pending records. Some implementations may recognize other special values.

params
A list of bindings, named or unnamed.