DBI methods

Implementations of pure virtual functions defined in the DBI package.

# S4 method for KazamConnection
dbBegin(conn, ...)

# S4 method for KazamResult
dbBind(res, params, ...)

# S4 method for KazamResult
dbClearResult(res, ...)

# S4 method for KazamResult
dbColumnInfo(res, ...)

# S4 method for KazamConnection
dbCommit(conn, ...)

# S4 method for KazamDriver
dbConnect(drv, ...)

# S4 method for KazamConnection,ANY
dbDataType(dbObj, obj, ...)

# S4 method for KazamDriver,ANY
dbDataType(dbObj, obj, ...)

# S4 method for KazamDriver,list
dbDataType(dbObj, obj, ...)

# S4 method for KazamConnection
dbDisconnect(conn, ...)

# S4 method for KazamConnection,character
dbExistsTable(conn, name, ...)

# S4 method for KazamResult
dbFetch(res, n = -1, ...)

# S4 method for KazamConnection
dbGetInfo(dbObj, ...)

# S4 method for KazamDriver
dbGetInfo(dbObj, ...)

# S4 method for KazamResult
dbGetInfo(dbObj, ...)

# S4 method for KazamResult
dbGetRowCount(res, ...)

# S4 method for KazamResult
dbGetRowsAffected(res, ...)

# S4 method for KazamResult
dbGetStatement(res, ...)

# S4 method for KazamResult
dbHasCompleted(res, ...)

# S4 method for KazamConnection
dbIsValid(dbObj, ...)

# S4 method for KazamDriver
dbIsValid(dbObj, ...)

# S4 method for KazamResult
dbIsValid(dbObj, ...)

# S4 method for KazamConnection,character
dbListFields(conn, name, ...)

# S4 method for KazamConnection
dbListTables(conn, ...)

# S4 method for KazamConnection,character
dbQuoteIdentifier(conn, x, ...)

# S4 method for KazamConnection,character
dbQuoteString(conn, x, ...)

# S4 method for KazamConnection,character
dbReadTable(conn, name, ...)

# S4 method for KazamConnection,character
dbRemoveTable(conn, name, ...)

# S4 method for KazamConnection
dbRollback(conn, ...)

# S4 method for KazamConnection,character
dbSendQuery(conn, statement, ...)

# S4 method for KazamConnection,character
dbSendStatement(conn, statement, ...)

# S4 method for KazamConnection,character,data.frame
dbWriteTable(conn, name, value, overwrite = FALSE, append = FALSE, ...)

# S4 method for KazamConnection
show(object)

# S4 method for KazamDriver
show(object)

# S4 method for KazamResult
show(object)

Arguments

conn

A DBIConnection object, as returned by dbConnect().

...

Other parameters passed on to methods.

res

An object inheriting from DBIResult.

params

A list of bindings, named or unnamed.

drv

an object that inherits from DBIDriver, or an existing DBIConnection object (in order to clone an existing connection).

dbObj

A object inheriting from DBIDriver or DBIConnection

obj

An R object whose SQL type we want to determine.

name

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')

n

maximum number of records to retrieve per fetch. Use n = -1 or n = Inf to retrieve all pending records. Some implementations may recognize other special values.

x

A character vector, SQL or Id object to quote as identifier.

statement

a character string containing SQL.

value

a data.frame (or coercible to data.frame).

overwrite

Allow overwriting the destination table. Cannot be TRUE if append is also TRUE.

append

Allow appending to the destination table. Cannot be TRUE if overwrite is also TRUE.

object

Any R object